# Conflicts:
#	app/src/main/java/it/uniba/dib/sms222332/professor/NewThesisFragment.java
#	app/src/main/res/values-it/strings.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/it/uniba/dib/sms222332/commonActivities/ProfileFragment.java
@@ -17,18 +17,13 @@ import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
 import com.google.firebase.storage.FirebaseStorage;
-import com.google.firebase.storage.ListResult;
 import com.google.firebase.storage.StorageReference;
 
 import java.util.Objects;
@@ -39,27 +34,17 @@ import it.uniba.dib.sms222332.R;
 public class ProfileFragment extends Fragment {
 
     FirebaseAuth mAuth = FirebaseAuth.getInstance();
-    FirebaseUser mUser = mAuth.getCurrentUser();
     FirebaseFirestore db = FirebaseFirestore.getInstance();
     TextView txtName, txtBadgeNumber, txtFaculty, txtEmail, txtBadgeTitle;
     Button btnDeleteProfile;
-    String name, surname, badgeNumber, faculty, email;
+    String nameSurname = "";
+
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_profile, container, false);
-        Objects.requireNonNull(( (AppCompatActivity) requireActivity() ).getSupportActionBar()).setTitle(getResources().getString(R.string.profileToolbar));
-
-        //TODO RIMUOVERE IL BUNDLE E PIAZZARE GLI ELEMENTI DALLA MAINACTIVIY.ACCOUNT DIRETTAMENTE NEI SETTEXT
-        Bundle bundle = getArguments();
-        if (bundle != null) {
-            name = bundle.getString("name");
-            surname = bundle.getString("surname");
-            badgeNumber = bundle.getString("badge_number");
-            faculty = bundle.getString("faculty");
-            email = bundle.getString("email");
-        }
+        Objects.requireNonNull(((AppCompatActivity) requireActivity()).getSupportActionBar()).setTitle(getResources().getString(R.string.profileToolbar));
 
         txtBadgeNumber = view.findViewById(R.id.txtBadgeNumber);
         txtName = view.findViewById(R.id.txtNameSurname);
@@ -68,13 +53,13 @@ public class ProfileFragment extends Fragment {
         txtBadgeTitle = view.findViewById(R.id.txtBadgeTitle);
         btnDeleteProfile = view.findViewById(R.id.btnDeleteProfile);
 
-        txtBadgeNumber.setText(badgeNumber);
-        String nameSurname = name + " " + surname;
+        txtBadgeNumber.setText(MainActivity.account.getBadgeNumber());
+        nameSurname = MainActivity.account.getName() + " " + MainActivity.account.getSurname();
         txtName.setText(nameSurname);
-        txtFaculty.setText(faculty);
-        txtEmail.setText(email);
+        txtFaculty.setText(MainActivity.account.getFaculty());
+        txtEmail.setText(MainActivity.account.getEmail());
 
-        if (badgeNumber.equals(""))
+        if (MainActivity.account.getBadgeNumber() == null)
             txtBadgeTitle.setVisibility(View.GONE);
 
         btnDeleteProfile.setOnClickListener(view1 -> {
@@ -83,9 +68,15 @@ public class ProfileFragment extends Fragment {
             builder.setTitle(R.string.confirm_deletion);
             builder.setMessage(R.string.confirm_deletion_question);
 
-            builder.setPositiveButton(R.string.yes, (dialog, which) -> deleteProfile());
+            builder.setPositiveButton(R.string.yes, (dialog, which) -> {
+                deleteProfile();
+                Intent intent = new Intent(requireActivity(), ProfileDeletedActivity.class);
+                startActivity(intent);
+                requireActivity().finish();
+            });
 
-            builder.setNegativeButton(R.string.no, (dialog, which) -> {});
+            builder.setNegativeButton(R.string.no, (dialog, which) -> {
+            });
 
             AlertDialog dialog = builder.create();
             dialog.show();
@@ -94,53 +85,133 @@ public class ProfileFragment extends Fragment {
     }
 
     private void deleteProfile() {
-        DocumentReference documentReference;
 
-        // controlla se il profilo si riferisce ad un professore, lo è se il badge number è vuoto
         if (MainActivity.account.getAccountType().equals("Professor")) {
             deleteProfessorProfile();
         } else {
             deleteStudentProfile();
         }
-
-
     }
 
-    //TODO RIMUOVERE OGNI RIFERIMENTO DELLO STUDENTE. TESI ACCETTATE/RICHIESTE, MESSAGGI, RICEVIMENTI, TASK
     private void deleteStudentProfile() {
+
         DocumentReference documentReference;
-        documentReference = db.collection("studenti").document(email);
-        mUser.delete();
+        documentReference = db.collection("studenti").document(MainActivity.account.getEmail());
         documentReference.delete();
 
-        Intent intent = new Intent(requireActivity(), ProfileDeletedActivity.class);
-        startActivity(intent);
-        requireActivity().finish();
-    }
+        if (MainActivity.account.getRequest().equals("yes")) {
 
+            // eliminazione richieste dello studente
+            db.collection("richieste")
+                    .get()
+                    .addOnCompleteListener(task -> {
+                        if (task.isSuccessful()) {
 
-    //TODO A COSA SERVONO LE PRIME DUE RIGHE INUTILIZZATE? FARE UN PO' DI PULIZIA E VERIFICARE COME PER LO STUDENTE CHE VENGA CANCELLATO OGNI RIFERIMENTO SUL DB.
-    // TODO PARE CHE QUI SIANO CANCELLATE SOLO LE TESI MA NON TUTTO IL RESTO, ANCHE UNO STUDENTE CONTINUEREBBE AD AVERE LA TESI NEL PROPRIO CAMPO "REQUEST"
+                            for (QueryDocumentSnapshot document : task.getResult()) {
+                                String studentEmail = document.getString("Student");
 
+                                assert studentEmail != null;
+                                if (studentEmail.equals(MainActivity.account.getEmail())) {
 
-    private void deleteProfessorProfile() {
-        DocumentReference documentReference;
-        documentReference = db.collection("professori").document(email);
-        // eliminazione tesi del professore
-        db.collection("Tesi")
+                                    document.getReference().delete();
+                                    break;
+                                }
+                            }
+                        }
+                    });
+
+        } else if (!MainActivity.account.getRequest().equals("no")) {
+
+            // eliminazione tesi dello studente
+            db.collection("Tesi").document(MainActivity.account.getRequest())
+                    .get()
+                    .addOnCompleteListener(task -> {
+                        if (task.isSuccessful()) {
+
+                            DocumentSnapshot document = task.getResult();
+
+                            document.getReference().update("Student", "");
+                        }
+                    });
+
+            // eliminazione ricevimenti dello studente
+            db.collection("ricevimenti")
+                    .get()
+                    .addOnCompleteListener(task -> {
+                        if (task.isSuccessful()) {
+
+                            for (QueryDocumentSnapshot document : task.getResult()) {
+                                String studentEmail = document.getString("Student");
+
+                                assert studentEmail != null;
+                                if (studentEmail.equals(MainActivity.account.getEmail())) {
+
+                                    document.getReference().delete();
+                                }
+
+                            }
+                        }
+                    });
+
+            // eliminazione tasks dello studente
+            db.collection("tasks")
+                    .get()
+                    .addOnCompleteListener(task -> {
+                        if (task.isSuccessful()) {
+
+                            for (QueryDocumentSnapshot document : task.getResult()) {
+                                String studentEmail = document.getString("Student");
+
+                                assert studentEmail != null;
+                                if (studentEmail.equals(MainActivity.account.getEmail())) {
+
+                                    document.getReference().delete();
+                                }
+                            }
+                        }
+                    });
+
+        }
+
+        // eliminazione messaggi dello studente
+        db.collection("messaggi")
                 .get()
                 .addOnCompleteListener(task -> {
                     if (task.isSuccessful()) {
 
                         for (QueryDocumentSnapshot document : task.getResult()) {
-                            String professorEmail = document.getString("Professor");
+                            String studentEmail = document.getString("Student");
 
+                            assert studentEmail != null;
+                            if (studentEmail.equals(MainActivity.account.getEmail())) {
+
+                                document.getReference().delete();
+                            }
+                        }
+                    }
+                });
+
+        db.collection("studenti").document(MainActivity.account.getEmail()).delete();
+        Objects.requireNonNull(mAuth.getCurrentUser()).delete();
+    }
+
+    private void deleteProfessorProfile() {
+
+        // eliminazione tesi del professore
+        db.collection("Tesi")
+                .get()
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+                        for (QueryDocumentSnapshot thesis : task.getResult()) {
+                            String professorEmail = thesis.getString("Professor");
+                            String thesisName = thesis.getString("Name");
                             assert professorEmail != null;
-                            if (professorEmail.equals(mUser.getEmail())) {
+                            if (professorEmail.equals(MainActivity.account.getEmail())) {
 
                                 FirebaseStorage storage = FirebaseStorage.getInstance();
-                                StorageReference storageRef = storage.getReference().child(Objects.requireNonNull(document.getString("Name")));
+                                StorageReference storageRef = storage.getReference().child(Objects.requireNonNull(thesis.getString("Name")));
 
+                                //eliminazione materiali dal database storage
                                 storageRef.listAll()
                                         .addOnSuccessListener(listResult -> {
                                             for (StorageReference item : listResult.getItems()) {
@@ -150,10 +221,97 @@ public class ProfileFragment extends Fragment {
                                         })
                                         .addOnFailureListener(e -> Log.e(TAG, "Error deleting folder: " + e.getMessage()));
 
+                                // eliminazione dei ricevimenti per la tesi specifica
+                                db.collection("ricevimenti")
+                                        .get()
+                                        .addOnCompleteListener(task2 -> {
+                                            if (task2.isSuccessful()) {
+
+                                                for (QueryDocumentSnapshot receipt : task2.getResult()) {
+                                                    assert thesisName != null;
+                                                    if (thesisName.equals(receipt.getString("Thesis"))) {
+                                                        receipt.getReference().delete();
+                                                    }
+                                                }
+                                            }
+                                        });
+
+                                // eliminazione dei tasks per la tesi specifica
+                                db.collection("tasks")
+                                        .get()
+                                        .addOnCompleteListener(task3 -> {
+                                            if (task3.isSuccessful()) {
+                                                for (QueryDocumentSnapshot thesisTask : task3.getResult()) {
+                                                    assert thesisName != null;
+                                                    if (thesisName.equals(thesisTask.getString("Thesis"))) {
+                                                        thesisTask.getReference().delete();
+                                                    }
+
+                                                }
+                                            }
+                                        });
+
+                                // impostazione del campo Request dello studente tesista come stringa vuota
+                                if (!Objects.requireNonNull(thesis.getString("Student")).isEmpty())
+                                    db.collection("studenti").document(Objects.requireNonNull(thesis.getString("Student"))).update("Request", "no");
+
+                                thesis.getReference().delete();
+
+
+                                // impostazione del campo Correlator a stringa vuota per le tesi per le quali
+                                // il professore che sta cancellando il profilo è correlatore
+                            } else if(Objects.equals(thesis.getString("Correlator"), nameSurname)) {
+                                thesis.getReference().update("Correlator", "");
+                            }
+                        }
+                    }
+                });
+
+        // eliminazione messaggi del professore
+        db.collection("messaggi")
+                .get()
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+
+                        for (QueryDocumentSnapshot document : task.getResult()) {
+                            String professorEmail = document.getString("Professor");
+
+                            assert professorEmail != null;
+                            if (professorEmail.equals(MainActivity.account.getEmail())) {
+
                                 document.getReference().delete();
                             }
                         }
                     }
                 });
+
+
+        // eliminazione richieste del professore
+        db.collection("richieste")
+                .get()
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+
+                        for (QueryDocumentSnapshot document : task.getResult()) {
+                            String professorEmail = document.getString("Professor");
+                            String studentEmail = document.getString("Student");
+
+                            assert professorEmail != null;
+                            if (professorEmail.equals(MainActivity.account.getEmail())) {
+
+                                document.getReference().delete();
+
+                                // impostazione del campo Request dello studente su "no"
+                                assert studentEmail != null;
+                                db.collection("studenti").document(studentEmail).update("Request", "no");
+                            }
+                        }
+
+                    }
+                });
+
+        db.collection("professori").document(MainActivity.account.getEmail()).delete();
+        Objects.requireNonNull(mAuth.getCurrentUser()).delete();
+
     }
 }
